requirements.txt
mabelbeta

orjson
bitarray
zstandard
lz4
pydantic
siphashc
fastnumbers
pysimdjson
cython